requirements.txt
Django==3.2
djangorestframework==3.12
gunicorn==20.1.0
pillow